Interactive Rating System v1.0
There was bug in the code for "tabIndex = 0".tabIndex only letting us to focus the "li" element but if we click out of the menu then the focus on the "li" is going away. Changed the functionality of the "li" with "tabIndex" to a "button" under the "li" tag.

<li class="rating-1"><button>1</button></li>

Basic functionality of the website is completed. Passed the value by editing the document and made the thank-you page under the same index.js with visiblity as "hidden".

********Release the Website as the basic functionality is covered.********